backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
celery==5.4.0
redis==5.0.0
nltk==3.8.1
numpy==1.26.0
pillow==10.0.0
opencv-python==4.8.0.76
spacy==3.7.0
langid==1.1.6
qdrant-client==1.9.0
google-generativeai==0.8.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.27.0
torch
torchvision
transformers
sentence-transformers
python-dotenv==1.0.0